Move TestDescriptors to inner classes
