+++ prototypes/javaspec-engine/src/main/java/info/javaspec/engine/JupiterSpec.java
@@ -2,6 +2,8 @@ package info.javaspec.engine;
 
 import info.javaspec.api.Verification;
 import org.junit.platform.engine.TestDescriptor;
+import org.junit.platform.engine.UniqueId;
+import org.junit.platform.engine.support.descriptor.AbstractTestDescriptor;
 
 public class JupiterSpec {
   private final String behavior;
@@ -20,5 +22,32 @@ public class JupiterSpec {
     );
 
     parentDescriptor.addChild(specDescriptor);
+  }
+
+  //TODO KDK: Does it make sense for this to be separate from JupiterSpec?
+  public static class SpecDescriptor extends AbstractTestDescriptor {
+    private final Verification verification;
+
+    public static SpecDescriptor forSpec(UniqueId parentId, String behavior, Verification verification) {
+      return new SpecDescriptor(
+        parentId.append("spec", behavior),
+        behavior,
+        verification
+      );
+    }
+
+    private SpecDescriptor(UniqueId uniqueId, String displayName, Verification verification) {
+      super(uniqueId, displayName);
+      this.verification = verification;
+    }
+
+    @Override
+    public Type getType() {
+      return Type.TEST;
+    }
+
+    public void runSpec() {
+      this.verification.execute();
+    }
   }
 }
